mac_requirements.txt
# ------------------------------------------------------------------------
# Python 3.10.16
# ------------------------------------------------------------------------

# ------------------------------------------------------------------------
# Use the CPU-only index for PyTorch and Torchaudio
# ------------------------------------------------------------------------
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.6.0
torchaudio>=2.6.0

# ------------------------------------------------------------------------
# Flask & Backend
# ------------------------------------------------------------------------
flask
flask-wtf
werkzeug
flask-login
flask-sqlalchemy
flask-bcrypt

# ------------------------------------------------------------------------
# Google Gemini API
# ------------------------------------------------------------------------

google-genai
google-auth

# ------------------------------------------------------------------------
# Whisper & Speech Processing
# ------------------------------------------------------------------------
openai-whisper
numpy==1.26.4
scipy

# ------------------------------------------------------------------------
# TTS (Text-to-Speech)
# ------------------------------------------------------------------------
edge-tts

# ------------------------------------------------------------------------
# Audio Processing
# ------------------------------------------------------------------------
moviepy==1.0.3
ffmpeg-python
pydub

# ------------------------------------------------------------------------
# PDF Processing
# NOTE: poppler-utils must be installed at the system level, not via pip.
# ------------------------------------------------------------------------
pdf2image
pillow==11.1.0
poppler-utils  # Required for pdf2image (Linux users)


# ------------------------------------------------------------------------
# kokoro
# ------------------------------------------------------------------------
#kokoro==0.8.4
#misaki[zh]==0.8.4

# ------------------------------------------------------------------------
# System Utilities
# ------------------------------------------------------------------------
tqdm
nest_asyncio
requests
python-dotenv
PyPDF2==3.0.1
ipython

transformers==4.30.0
pysoundfile